// frontend/src/api/generated/models/user-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UserRole } from './user-role';

/**
 * 
 * @export
 * @interface UserDetail
 */
export interface UserDetail {
    /**
     * ユーザーID
     * @type {number}
     * @memberof UserDetail
     */
    'id': number;
    /**
     * ログインID
     * @type {string}
     * @memberof UserDetail
     */
    'userCode': string;
    /**
     * 氏名
     * @type {string}
     * @memberof UserDetail
     */
    'fullName': string;
    /**
     * メールアドレス
     * @type {string}
     * @memberof UserDetail
     */
    'email': string;
    /**
     * 
     * @type {UserRole}
     * @memberof UserDetail
     */
    'role': UserRole;
    /**
     * 有効フラグ
     * @type {boolean}
     * @memberof UserDetail
     */
    'isActive': boolean;
    /**
     * アカウントロックフラグ
     * @type {boolean}
     * @memberof UserDetail
     */
    'locked': boolean;
    /**
     * ロック発生日時（未ロック時はnull）
     * @type {string}
     * @memberof UserDetail
     */
    'lockedAt'?: string | null;
    /**
     * パスワード変更要求フラグ
     * @type {boolean}
     * @memberof UserDetail
     */
    'passwordChangeRequired': boolean;
    /**
     * 登録日時
     * @type {string}
     * @memberof UserDetail
     */
    'createdAt': string;
    /**
     * 最終更新日時
     * @type {string}
     * @memberof UserDetail
     */
    'updatedAt': string;
    /**
     * 楽観的ロックバージョン
     * @type {number}
     * @memberof UserDetail
     */
    'version': number;
}